[KB] chore/[#137]: tsx 문법 안쓰는 소스 파일들 TypeScript파일로 바꾸기
