+++ src/api/havruta/havrutaBoard.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
-import { HavrutaBoard } from '~/models/Havruta';
-import { client } from '~/api/client';
-import { authClient } from '~/api/auth/authClient';
+import { HavrutaBoard } from '~/models/Havruta.ts';
+import { client } from '~/api/client.ts';
+import { authClient } from '~/api/auth/authClient.ts';
 
 // 페이지네이션 적용된 하브루타 게시물 모두 가져오기
 export const getHavrutaBoards = async (
