+++ src/models/Auth.ts
