+++ src/store/authStore.ts
@@ -13,7 +13,6 @@ import {
   signUp as signUpApi,
   reissueToken as reissueTokenApi,
 } from '~/api/auth/authApi';
-import { stringify } from 'querystring';
 
 // Zustand에서 관리할 상태의 구조, 데이터 Type 정의
 interface authStore {
